requirements.txt
kafka-python==2.0.2
streamlit==1.32.0
pandas==2.2.0
